Refactor KPI menu
I've moved from a set of useState methods to a single useReducer
so that actions etc can be specificed as the statemachine that
this element represents has started to grow and I will need to
use that state across a number of KPI objects.

I also fixed up the styles of the element so that the toggle
button text is smaller than the title and improved the spacing.