Implement reporting delay KPI
This encapsulates factoring out the KPI view information into a
separate component to use for both implemented KPIs. It has
exposed some issues with regards to how KPI data will be passed in
to the KeyPerformanceIndicators component. It has also raised an
issue around displaying different types of table + ations for
different KPIs.

For example, the "Number of Cases" KPI displays a set of data
over N months with one column per month. It feels like the date
ranges that a user will need to view data for should be custom but
limited to N months where N won't break the layout.

On the other hand the "Reporting Delay" KPI shows traditional
aggreate data for a specific set of groupings of data. Not only
do we want to show a custom table cell for percentages but we
also may want to offer different date range selections. This is
because the space required to show data is not related to the date
range we select data over.

I suspect we'll need a separate set of files for each KPI.