Fix completed case action plans
Previously, we checked `#present?` on the contents of fields to see if
they contained a value. Unfortuantely, `#present?` returns false on the
false value as false is not considered a value which is present. We use
false as a value to some answered fields, which implies that false is
infact a present value for us. By using `#nil?` we side step this issue
BUT this assumes that we don't use any other forms of empty value for
form fields e.g. "", [], {}, etc as these will be falsly reported as
completed when they are not :(.